+++ pyam/core.py
@@ -50,6 +50,7 @@ from pyam.utils import (
 from pyam.read_ixmp import read_ix
 from pyam.timeseries import fill_series
 from pyam._aggregate import _aggregate, _aggregate_region, _group_and_agg
+from pyam.units import convert_unit, convert_unit_with_mapping
 
 logger = logging.getLogger(__name__)
 
@@ -251,7 +252,7 @@ class IamDataFrame(object):
         if self.time_col is not other.time_col:
             raise ValueError('incompatible time format (years vs. datetime)!')
 
-        ret = copy.deepcopy(self) if not inplace else self
+        ret = self.copy() if not inplace else self
 
         diff = other.meta.index.difference(ret.meta.index)
         intersect = other.meta.index.intersection(ret.meta.index)
@@ -718,7 +719,7 @@ class IamDataFrame(object):
             return self.append(df.rename(mapping), inplace=inplace)
 
         # if append is False, iterate over rename mapping and do groupby
-        ret = copy.deepcopy(self) if not inplace else self
+        ret = self.copy() if not inplace else self
 
         # renaming is only applied where a filter matches for all given columns
         rows = ret._apply_filters(**filters)
@@ -757,26 +758,42 @@ class IamDataFrame(object):
         if not inplace:
             return ret
 
-    def convert_unit(self, conversion_mapping, inplace=False):
-        """Converts units based on provided unit conversion factors
+    def convert_unit(self, current, to=None, factor=None, inplace=False):
+        """Converts a unit using a given factor or the pint package
+
+        The `pint package <https://pint.readthedocs.io>`_ natively handles
+        conversion of standard (SI) units (e.g., exajoule to terawatt-hours,
+        :code:`EJ -> TWh`). It can also parse combined units (e.g.,
+        exajoule per year, :code:`EJ/yr`).
+
+        The :py:class:`pint.UnitRegistry` used by default loads additional
+        unit definitions relevant for integrated assessment models and energy
+        systems analysis from the `IAMconsortium/units
+        <https://github.com/IAMconsortium/units>`_ repository.
+
+        You can access the :py:class:`pint.UnitRegistry` used by :class:`pyam`
+        via :func:`pint.get_application_registry`.
 
         Parameters
         ----------
-        conversion_mapping: dict
-            for each unit for which a conversion should be carried out,
-            provide current unit and target unit and conversion factor
-            {<current unit>: [<target unit>, <conversion factor>]}
+        current: str (or mapping, deprecated)
+            name of current unit (to be converted from)
+        to: str
+            name of new unit (to be converted to)
+        factor: value, optional
+            conversion factor if given, otherwise defaults to the application
+            :py:class:`pint.UnitRegistry`
         inplace: bool, default False
             if True, do operation inplace and return None
         """
-        ret = copy.deepcopy(self) if not inplace else self
-        for current_unit, (new_unit, factor) in conversion_mapping.items():
-            factor = pd.to_numeric(factor)
-            where = ret.data['unit'] == current_unit
-            ret.data.loc[where, 'value'] *= factor
-            ret.data.loc[where, 'unit'] = new_unit
-        if not inplace:
-            return ret
+        # TODO: deprecate using `dict` in next release (>=0.6.0)
+        # TODO: make `to` required
+        if isinstance(current, dict) and to is None and factor is None:
+            deprecation_warning('Use explicit keyword arguments instead!',
+                                type='Using a dictionary to convert units')
+            return convert_unit_with_mapping(self, current, inplace)
+        # new standard method, remove this comment when deprecating above
+        return convert_unit(self, current, to, factor, inplace)
 
     def normalize(self, inplace=False, **kwargs):
         """Normalize data to a specific data point
@@ -791,7 +808,7 @@ class IamDataFrame(object):
         """
         if len(kwargs) > 1 or self.time_col not in kwargs:
             raise ValueError('Only time(year)-based normalization supported')
-        ret = copy.deepcopy(self) if not inplace else self
+        ret = self.copy() if not inplace else self
         df = ret.data
         # change all below if supporting more in the future
         cols = self.time_col
@@ -1120,7 +1137,7 @@ class IamDataFrame(object):
 
         _keep = self._apply_filters(**kwargs)
         _keep = _keep if keep else ~_keep
-        ret = copy.deepcopy(self) if not inplace else self
+        ret = self.copy() if not inplace else self
         ret.data = ret.data[_keep]
 
         idx = _make_index(ret.data)
@@ -1513,7 +1530,7 @@ class IamDataFrame(object):
         mapping = read_pandas(fname).rename(str.lower, axis='columns')
         map_col = map_col.lower()
 
-        ret = copy.deepcopy(self) if not inplace else self
+        ret = self.copy() if not inplace else self
         _df = ret.data
         columns_orderd = _df.columns
 
@@ -1813,7 +1830,7 @@ def concat(dfs):
     for df in dfs:
         df = df if isinstance(df, IamDataFrame) else IamDataFrame(df)
         if _df is None:
-            _df = copy.deepcopy(df)
+            _df = df.copy()
         else:
             _df.append(df, inplace=True)
     return _df