+++ tests/test_feature_compare.py
@@ -1,6 +1,5 @@
 import datetime as dt
 
-
 import numpy as np
 import pandas as pd
 from pyam import compare, IAMC_IDX
@@ -15,11 +14,11 @@ def test_compare(test_df):
     obs = compare(test_df, clone, right_label='test_df', left_label='clone')
 
     exp = pd.DataFrame([
-        ['Primary Energy', 'EJ/y', dt.datetime(2005, 6, 17), 2, 1],
-        ['Primary Energy|Coal', 'EJ/y', dt.datetime(2005, 6, 17), np.nan, 0.5],
-        ['Primary Energy|Coal', 'EJ/y', dt.datetime(2010, 7, 21), np.nan, 3],
-        ['Primary Energy|Gas', 'EJ/y', dt.datetime(2005, 6, 17), 0.5, np.nan],
-        ['Primary Energy|Gas', 'EJ/y', dt.datetime(2010, 7, 21), 3, np.nan],
+        ['Primary Energy', 'EJ/yr', dt.datetime(2005, 6, 17), 2, 1],
+        ['Primary Energy|Coal', 'EJ/yr', dt.datetime(2005, 6, 17), np.nan, .5],
+        ['Primary Energy|Coal', 'EJ/yr', dt.datetime(2010, 7, 21), np.nan, 3],
+        ['Primary Energy|Gas', 'EJ/yr', dt.datetime(2005, 6, 17), 0.5, np.nan],
+        ['Primary Energy|Gas', 'EJ/yr', dt.datetime(2010, 7, 21), 3, np.nan],
     ],
         columns=['variable', 'unit', 'time', 'test_df', 'clone'],
     )